Fixed unsafe type cast warnings in Profiler and ProfilerClient modules.
#jira UE-166275
#rb Catalin.Dragoiu
#preflight 6352623e777a77c4403df7ee

[CL 22686686 by ionut matasaru in ue5-main branch]

+++ Engine/Source/Developer/Profiler/Private/ProfilerDataProvider.cpp
@@ -51,7 +51,7 @@ const uint32 IDataProvider::AdvanceFrame(const float DeltaTimeMS)
 	FrameIndices.Add(FIntPoint(FrameIndex, SampleEndIndex));
 
 	FrameTimes.Add(DeltaTimeMS);
-	ElapsedFrameTimes.Add(ElapsedTimeMS);
+	ElapsedFrameTimes.Add(static_cast<float>(ElapsedTimeMS));
 
 	// Update the values.
 	FrameCounters.Last() = LastSecondFrameCounter;
@@ -120,9 +120,9 @@ const uint32 FArrayDataProvider::AddHierarchicalSample(
 	if(FProfilerSample::IsIndexValid(InParentIndex))
 	{
 		FProfilerSample& Parent = Collection[ InParentIndex ];
-		const uint32 InitialMemUsage = Parent.ChildrenIndices().GetAllocatedSize();
+		const SIZE_T InitialMemUsage = Parent.ChildrenIndices().GetAllocatedSize();
 		Parent.AddChild(HierSampleIndex);
-		const uint32 AccumulatedMemoryUsage = Parent.ChildrenIndices().GetAllocatedSize() - InitialMemUsage;
+		const SIZE_T AccumulatedMemoryUsage = Parent.ChildrenIndices().GetAllocatedSize() - InitialMemUsage;
 		ChildrenIndicesMemoryUsage += AccumulatedMemoryUsage;
 	}
