+++ Engine/Source/Developer/Profiler/Private/ProfilerSession.cpp
@@ -460,7 +460,7 @@ bool FProfilerSession::HandleTicker( float DeltaTime )
 					ThreadDuration.DurationCycles, 1, 
 					FrameRootSampleIndex 
 				);
-				ThreadMS.FindOrAdd( ThreadID ) = StatMetaData->ConvertCyclesToMS( ThreadDuration.DurationCycles );
+				ThreadMS.FindOrAdd( ThreadID ) = static_cast<float>(StatMetaData->ConvertCyclesToMS( ThreadDuration.DurationCycles ));
 
 				// Recursively add children and parent to the root samples.
 				for( int32 Index = 0; Index < ThreadGraph.Children.Num(); Index++ )
@@ -486,7 +486,7 @@ bool FProfilerSession::HandleTicker( float DeltaTime )
 		MutableCollection[FrameRootSampleIndex].SetDurationCycles( GameThreadCycles != 0 ? GameThreadCycles : MaxThreadCycles );
 
 		// Update FPS analyzer.
-		const float GameThreadTimeMS = StatMetaData->ConvertCyclesToMS( GameThreadCycles );
+		const float GameThreadTimeMS = static_cast<float>(StatMetaData->ConvertCyclesToMS( GameThreadCycles ));
 		FPSAnalyzer->AddSample( GameThreadTimeMS > 0.0f ? 1000.0f / GameThreadTimeMS : 0.0f );
 
 		// Process the non-hierarchical samples for the specified frame.
@@ -509,7 +509,7 @@ bool FProfilerSession::HandleTicker( float DeltaTime )
 
 		// Advance frame
 		const uint32 DataProviderFrameIndex = DataProvider->GetNumFrames();
-		DataProvider->AdvanceFrame( StatMetaData->ConvertCyclesToMS( MaxThreadCycles ) );
+		DataProvider->AdvanceFrame( static_cast<float>(StatMetaData->ConvertCyclesToMS( MaxThreadCycles )) );
 
 		// Update aggregated stats
 		UpdateAggregatedStats( DataProviderFrameIndex );
@@ -948,17 +948,17 @@ void FProfilerAggregatedStat::Advance()
 		}
 
 		{
-			_MinValueAllFrames = FMath::Min<float>( _MinValueAllFrames, _ValueOneFrame );
-			_MaxValueAllFrames = FMath::Max<float>( _MaxValueAllFrames, _ValueOneFrame );
+			_MinValueAllFrames = FMath::Min<double>( _MinValueAllFrames, _ValueOneFrame );
+			_MaxValueAllFrames = FMath::Max<double>( _MaxValueAllFrames, _ValueOneFrame );
 
-			_MinNumCallsAllFrames = FMath::Min<float>( _MinNumCallsAllFrames, _NumCallsOneFrame );
-			_MaxNumCallsAllFrames = FMath::Max<float>( _MaxNumCallsAllFrames, _NumCallsOneFrame );
+			_MinNumCallsAllFrames = FMath::Min<uint32>( _MinNumCallsAllFrames, _NumCallsOneFrame );
+			_MaxNumCallsAllFrames = FMath::Max<uint32>( _MaxNumCallsAllFrames, _NumCallsOneFrame );
 		}
 	}
 	else
 	{
-		_MinValueAllFrames = FMath::Min<float>( _MinValueAllFrames, _ValueOneFrame );
-		_MaxValueAllFrames = FMath::Max<float>( _MaxValueAllFrames, _ValueOneFrame );
+		_MinValueAllFrames = FMath::Min<double>( _MinValueAllFrames, _ValueOneFrame );
+		_MaxValueAllFrames = FMath::Max<double>( _MaxValueAllFrames, _ValueOneFrame );
 	}
 
 	_ValueOneFrame = 0.0f;
